+++ RELEASE_NOTES.md
@@ -1,11 +1,16 @@
 Authentication Service MKII release notes
 =========================================
 
+0.2.5
+-----
+* OrcID is now supported as an identity provider. See `deploy.cfg.example` for a
+  configuration example.
+
 0.2.4
 -----
 * CONFIGURATION CHANGE - the templates directory is now configurable. Add
   `template-dir = templates` to any existing configuration files to preserve current
-  behavior. This purpose of this change is primarily to allow other applications to
+  behavior. The purpose of this change is primarily to allow other applications to
   test the server with the templates in a non-standard location.
 
 0.2.3